Instantiatable connectors and file based ServerConfiguration
Enable non-static connector instances to read from different lms
systems. Use ServerConfiguration to read from xml config files and
remove old hard coded config

+++ src/main/java/de/lemo/dms/service/ServiceLoadWholeDatabase.java
@@ -21,7 +21,7 @@ public class ServiceLoadWholeDatabase extends BaseService {
 
     @GET
     @Produces(MediaType.APPLICATION_JSON)
-    public JSONObject loadWholeDatabase(@QueryParam("connector") Integer connectorId) {
+    public JSONObject loadWholeDatabase(@QueryParam("connector") Long connectorId) {
         super.logger.info("call for service: loadWholeDatabase");
 
         ConnectorManager cm = ConnectorManager.getInstance();
